+++ Web/src/main/java/com/cy/homeManagement/house_unit/controller/HouseUnitController.java
@@ -3,12 +3,16 @@ package com.cy.homeManagement.house_unit.controller;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.cy.CommonResult;
+import com.cy.homeManagement.house_list.entity.HouseList;
+import com.cy.homeManagement.house_list.mapper.HouseListMapper;
 import com.cy.homeManagement.house_unit.entity.HouseUnit;
 import com.cy.homeManagement.house_unit.entity.HouseUnitParam;
 import com.cy.homeManagement.house_unit.serveie.HouseUnitService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import javax.annotation.Resource;
 import javax.validation.Valid;
 import java.util.List;
 
@@ -23,6 +27,8 @@ import java.util.List;
 public class HouseUnitController {
     @Autowired
     private HouseUnitService houseUnitService;
+    @Resource
+    private HouseListMapper houseListMapper;
 
     /**
      * 房屋单元列表
@@ -38,6 +44,7 @@ public class HouseUnitController {
      * @param houseUnit
      * @return
      */
+    @PreAuthorize("hasAuthority('sys:houseUnit:add')")
     @PostMapping
     public CommonResult<String> addHouseUnit(@RequestBody @Valid HouseUnit houseUnit){
 
@@ -61,20 +68,24 @@ public class HouseUnitController {
      * @param houseUnit
      * @return
      */
+    @PreAuthorize("hasAuthority('sys:houseUnit:edit')")
     @PutMapping
     public CommonResult<String> editHouseUnit(@RequestBody @Valid  HouseUnit houseUnit){
         QueryWrapper<HouseUnit> queryWrapper = new QueryWrapper<>();
-        queryWrapper.lambda().eq(HouseUnit::getUnitName,houseUnit.getUnitName());
-        HouseUnit one = houseUnitService.getOne(queryWrapper);
-        if (one != null){
-            return CommonResult.error("单元已存在!");
-        }else {
+        queryWrapper.lambda().eq(HouseUnit::getBuildId,houseUnit.getBuildId());
+        List<HouseUnit> list = houseUnitService.list(queryWrapper);
+        for (HouseUnit houseUnit1:list) {
+            if (houseUnit1.getUnitName().equals(houseUnit.getUnitName())){
+                return CommonResult.error("单元已存在!");
+            }
+        }
+
             boolean editStatus = houseUnitService.updateById(houseUnit);
             if (editStatus) {
                 return CommonResult.success("编辑单元成功!");
             }
             return CommonResult.error("编辑单元失败!");
-        }
+
     }
 
     /**
@@ -82,13 +93,21 @@ public class HouseUnitController {
      * @param unitId
      * @return
      */
+    @PreAuthorize("hasAuthority('sys:houseUnit:delete')")
     @DeleteMapping("/{unitId}")
     public CommonResult<String> deleteHouseUnit(@PathVariable("unitId") @Valid  Long unitId){
-        boolean deleteStatus = houseUnitService.removeById(unitId);
-        if(deleteStatus){
-            return CommonResult.success("删除单元成功!");
+        QueryWrapper<HouseList> houseListQueryWrapper = new QueryWrapper<>();
+        houseListQueryWrapper.lambda().eq(HouseList::getHouseId,unitId);
+        List<HouseList> liveHouses = houseListMapper.selectList(houseListQueryWrapper);
+        if (liveHouses.size() > 0){
+            return CommonResult.error("该单元已有房屋，不能删除");
+        }else {
+            boolean deleteStatus = houseUnitService.removeById(unitId);
+            if (deleteStatus) {
+                return CommonResult.success("删除单元成功!");
+            }
+            return CommonResult.error("删除单元失败!");
         }
-        return CommonResult.error("删除单元失败!");
     }
 
     /**
